+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 6.0.0 (unreleased)
 
 - Dropped support for Elasticsearch 7 and OpenSearch 1
+- Dropped support for Active Record < 7.2
 
 ## 5.5.2 (2025-05-20)