+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## 6.0.0 (unreleased)
 
 - Added support for `exists: false`
-- Removed quantization for `knn` option for Elasticsearch 8.14+
+- Added `quantization` to `knn` option for Elasticsearch
+- Removed default quantization for `knn` option for Elasticsearch 8.14+
 - Removed `execute` option (no longer needed)
 - Dropped support for Elasticsearch 7 and OpenSearch 1
 - Dropped support for Active Record < 7.2